+++ aliases-mac.zsh
@@ -1,5 +1,3 @@
-alias c="pbcopy"
-alias p="pbpaste"
 alias cat="ccat"
 alias nn="nvim"
 alias np="nvim --clean -n -c \"set nobackup noswapfile noundofile\""